set token in project action

+++ src/actions/project/projectAction.ts
@@ -1,19 +1,28 @@
 "use server";
+import { getAdminData } from "@/helpers/getAdminData";
 import { IProjectFormData } from "@/types";
 import { revalidateTag } from "next/cache";
+import { cookies } from "next/headers";
 
 // ---------------- CREATE ----------------
 export const createProjectServerAction = async (
   projectData: IProjectFormData
 ) => {
+  const cookieStore = await cookies();
+  const token = cookieStore.get("accessToken")?.value;
+  if (!token) throw new Error("No token found");
+
+  const admin = await getAdminData();
+
   const payload = {
     ...projectData,
-    ownerId: 1,
+    ownerId: admin?.id,
   };
   const res = await fetch(`${process.env.NEXT_PUBLIC_BASE_API}/project`, {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
+      Authorization: `Bearer ${token}`,
     },
     body: JSON.stringify(payload),
     // credentials: "include",
@@ -28,12 +37,17 @@ export const updateProjectServerAction = async (
   slug: string,
   updateProjectData: IProjectFormData
 ) => {
+  const cookieStore = await cookies();
+  const token = cookieStore.get("accessToken")?.value;
+  if (!token) throw new Error("No token found");
+
   const res = await fetch(
     `${process.env.NEXT_PUBLIC_BASE_API}/project/${slug}`,
     {
       method: "PATCH",
       headers: {
         "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
       },
       body: JSON.stringify(updateProjectData),
       // credentials: "include",
@@ -47,12 +61,17 @@ export const updateProjectServerAction = async (
 
 // ---------------- DELETE ----------------
 export const deleteProjectServerAction = async (slug: string) => {
+  const cookieStore = await cookies();
+  const token = cookieStore.get("accessToken")?.value;
+  if (!token) throw new Error("No token found");
+
   const res = await fetch(
     `${process.env.NEXT_PUBLIC_BASE_API}/project/${slug}`,
     {
       method: "DELETE",
       headers: {
         "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
       },
       // credentials: "include",
     }
